# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Horror Movie Reviews API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Horror Movie Reviews application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Review": {
        "type": "object",
        "required": [
          "id",
          "authorId",
          "title",
          "content",
          "rating"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for a review",
            "example": "review_abc123"
          },
          "authorId": {
            "type": "string",
            "description": "The unique identifier for a user",
            "example": "1"
          },
          "title": {
            "type": "string",
            "description": "The title of the review",
            "example": "The best movie of all time: Test"
          },
          "content": {
            "type": "string",
            "description": "The content of the review",
            "example": "I love this movie so much"
          },
          "rating": {
            "type": "number",
            "description": "The rating of the movie for the review",
            "example": 10
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/reviews": {
      "get": {
        "summary": "Retrieve a list of reviews with optional filtering",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Maximum number of reviews to return"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "rating"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "example": "Test Movie"
                  },
                  "content": {
                    "type": "string",
                    "example": "I love this movie so much"
                  },
                  "rating": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/reviews/id": {
      "get": {
        "summary": "Retrieve a review by specified id",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the review"
          }
        ],
        "responses": {
          "200": {
            "description": "A review",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific review's information",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the review"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "comment"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "example": "Test Movie"
                  },
                  "content": {
                    "type": "string",
                    "example": "I love this movie so much"
                  },
                  "rating": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to update this review"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a Review by specified id",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the Review"
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}